+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Checkpoint] - 2026-01-07 09:50:28
+
+### Changed Files
+- `supabase/functions/instagram-oauth-callback/index.ts`
+
+### Statistics
+```
+ 1 file changed, 3 insertions(+), 3 deletions(-)
+```
+
 ## [Checkpoint] - 2026-01-07 09:20:42
 
 ### Changed Files